revise code in comments/routes.py and question/routes.py

+++ app/comments/routes.py
@@ -21,8 +21,8 @@ def add_comment():
     user_id = request.args.get('user_id')
     question_id = request.args.get('question_id')
     content = request.args.get('content')
-    date_created = request.args.get('date_created')
-    date_updated = request.args.get('date_updated')
+    date_created = datetime.now()
+    date_updated = datetime.now()
     is_anon = int(request.args.get('is_anonymous'))
     source = ('source')
 
@@ -42,15 +42,17 @@ def add_comment():
                                 is_anonymous=is_anon,
                                 source=source)
         db.session.add(new_comment) # adds a new comment to the database
-        db.session.commit() # commit all changes to the database
 
-        ###TODO
-        # need to add 1 to "num_comments" column of that question        
+        # need to add 1 to "num_comments" column of that question
+        question = Question.query.filter_by(question_id=question_id).first()
+        old_num = int(question.num_comments)
+        question.num_comments = old_num + 1
         
+        db.session.commit() # commit all changes to the database
         return make_response(f"Comment successfully created!", 200)
 
+    # if there is missing information
     return make_response(f"Unable to create a comment due to missing information!", 400) 
-    # print("TODO")
 
 
 @comments_bp.route('/get_comments', methods=['GET'])
@@ -61,6 +63,7 @@ def get_comments():
    """
    # get query parameters
    question_id = request.args.get('question_id')
+
    # check if question_is is None or not
    if question_id is not None:
        # get all comments that is from the question_id
@@ -75,7 +78,10 @@ def get_comments():
                "date_updated": comment.date_updated,
                "is_anon": comment.is_anonymous,
                "source": comment.source})
+       
        return jsonify(response)
+   
+   # if question_id is None
    return make_response(f"Unable to return comments due to missing question_id!", 400)
 
 
@@ -84,49 +90,56 @@ def update_comment():
     """Update the comment with id 'comment_id' that is associated with the
     question with id 'question_id'.
 
-    Not sure if we need question_id or not?
     """
 
     # get query parameters
-    # question_id = request.args.get('question_id')
     comment_id = request.args.get('comment_id')
     comment_new = request.args.get('content')    
     is_anon = int(request.args.get('is_anonymous'))
 
-    # query the comment that need to be updated
-    update = Comment.query.filter_by(comment_id=comment_id).first()
-    print("update: " + str(update))
-    # update the following fields for the comment
-    update.content = comment_new
-    print("current time: " + datetime.now())
-    update.date_updated = datetime.now()
-    update.is_anonymous = is_anon
+    # check if there is any missing variable
+    if comment_id is not None and comment_new is not None and is_anon is not None:
+        # query the comment that need to be updated
+        update = Comment.query.filter_by(comment_id=comment_id).first()
+        # update the following fields for the comment
+        update.content = comment_new
+        update.date_updated = datetime.now()
+        update.is_anonymous = is_anon
 
-    db.session.commit()
-    return make_response(f"Comment successfully updated!", 200)
+        db.session.commit()
+        return make_response(f"Comment successfully updated!", 200)
+    
+    # if one of the variables is missing
+    return make_response(f"Unable to update comments due to missing information!", 400)
 
 
 @comments_bp.route('/delete_comment', methods=['DELETE'])
 def delete_comment():
     """Delete the comment with id 'comment_id' that is associated with the
     question with id 'question_id'.
+    subtract 1 from the num_comments column for that question
+
     """
 
     # get query parameters
     question_id = request.args.get('question_id')
     comment_id = request.args.get('comment_id')
 
-    # query the comment with the id of comment_id
-    delete_com = Comment.query.filter_by(comment_id=comment_id).first()
-    # delete the comment from the database
-    db.session.delete(delete_com)
+    # check if there is any missing variable
+    if question_id is not None and comment_id is not None:
+        # query the comment with the id of comment_id
+        delete_com = Comment.query.filter_by(comment_id=comment_id).first()
+        # delete the comment from the database
+        db.session.delete(delete_com)
 
-    # num_comments in the question table need to subtract 1
-    question = Question.query.filter_by(question_id=question_id).first()
-    old_num = question.num_comments # get the current num_comments
-    # subtract by 1 and update the num_comments
-    question.num_comments = old_num - 1
+        # num_comments in the question table need to subtract 1
+        question = Question.query.filter_by(question_id=question_id).first()
+        old_num = int(question.num_comments) # get the current num_comments
+        # subtract by 1 and update the num_comments
+        question.num_comments = old_num - 1
+        
+        db.session.commit()
+        return make_response(f"Comment successfully deleted!", 200) 
     
-    db.session.commit()
-    
-    return make_response(f"Comment successfully deleted!", 200)    
+    # if there is missing variable
+    return make_response(f"Unable to delete the comment due to missing information!", 400)
